+++ libyul/backends/evm/AsmCodeGen.cpp
@@ -117,24 +117,6 @@ void EthAssemblyAdapter::appendJumpToIf(LabelID _labelId, JumpType _jumpType)
 	appendJumpInstruction(evmasm::Instruction::JUMPI, _jumpType);
 }
 
-void EthAssemblyAdapter::appendBeginsub(LabelID, int)
-{
-	// TODO we could emulate that, though
-	yulAssert(false, "BEGINSUB not implemented for EVM 1.0");
-}
-
-void EthAssemblyAdapter::appendJumpsub(LabelID, int, int)
-{
-	// TODO we could emulate that, though
-	yulAssert(false, "JUMPSUB not implemented for EVM 1.0");
-}
-
-void EthAssemblyAdapter::appendReturnsub(int, int)
-{
-	// TODO we could emulate that, though
-	yulAssert(false, "RETURNSUB not implemented for EVM 1.0");
-}
-
 void EthAssemblyAdapter::appendAssemblySize()
 {
 	m_assembly.appendProgramSize();
@@ -239,7 +221,6 @@ void CodeGenerator::assemble(
 		EVMDialect::strictAssemblyForEVM(_evmVersion),
 		builtinContext,
 		_optimizeStackAllocation,
-		false,
 		_identifierAccess,
 		_useNamedLabelsForFunctions
 	);